add retrying work queue periodically
scrobbles in the work queue are now retried every 10 seconds. the current
problem is that it will continue trying the queue even if there is no
work and that "every 10 seconds" is probably not the best strategy for
retries. potentially implement something that tries at increasingly
spaced out intervals up to some cap.