fixed work queue bug
previously we were writing the queue to disk for the first time after
waiting for the client to be initialized. this led to the queue file
being corrupted if the program panicked during client initialization
(which it currently does 100% of the time). now write() is no longer
public and the external user of the work queue doesnt have to worry
about taking care of it at the correct time because the queue does it
automatically.